fix(skb):优化符号显示效果

+++ src/main/java/com/yuyan/imemodule/adapter/CandidatesAdapter.kt
@@ -56,7 +56,7 @@ class CandidatesAdapter(context: Context?) :
         init {
             textView = view.findViewById(R.id.gv_item)
             textView.setTextColor(textColor)
-            textView.textSize = DevicesUtils.px2dip(mCandidateTextSize.toFloat()).toFloat()
+            textView.textSize = DevicesUtils.px2dip(mCandidateTextSize)
         }
     }
 }
